# storegateway: Reduce number of overlapping iterator interfaces into single iterator[S]