#### What this PR does

While working through the `storegateway` code, I found myself being confused with how different parts rely on different "iterators", which all have somewhere similar names, e.g. `seriesChunkRefsSetIterator`, `seriesChunkRefsIterator`, and all implement the same (already defined) interface.

I understand the historic aspects of that, but since the package already has `genericIterator[S]` in some places, I wanted to clean this up a bit, and converge all iterator interfaces into a single `iterator[S]`.

This PR is a very mechanical refactor of the package, which doesn't change any of behaviours of the code.

cc @dimitarvdimitrov

#### Which issue(s) this PR fixes or relates to

n/a

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.