@narqo 2024-02-23
I haven't noticed any particular regression or improvements in the package's benchmarks overall. I think, the ones where these µ-changes could be noticeable are the `BenchmarkMergedSeriesChunkRefsSetIterators` (they aren't dominated by any IO). There the changes in the results are within 0.1-0.3% (plus or minus, depending on when I run them :)

```
› benchstat ~/tmp/{1before,2after}.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/storegateway
                                                                                           │ /Users/v/tmp/1before.txt │      /Users/v/tmp/2after.txt       │
                                                                                           │          sec/op          │   sec/op     vs base               │
MergedSeriesChunkRefsSetIterators/with_duplicated_series_=_true_number_of_iterators_=_4-4                 11.65µ ± 0%   11.57µ ± 1%  -0.66% (p=0.025 n=10)
MergedSeriesChunkRefsSetIterators/with_duplicated_series_=_false_number_of_iterators_=_4-4                20.24µ ± 1%   20.27µ ± 0%       ~ (p=0.565 n=10)
geomean                                                                                                   15.36µ        15.32µ       -0.26%

                                                                                           │ /Users/v/tmp/1before.txt │       /Users/v/tmp/2after.txt       │
                                                                                           │           B/op           │     B/op      vs base               │
MergedSeriesChunkRefsSetIterators/with_duplicated_series_=_true_number_of_iterators_=_4-4                1.539Ki ± 0%   1.542Ki ± 0%       ~ (p=0.302 n=10)
MergedSeriesChunkRefsSetIterators/with_duplicated_series_=_false_number_of_iterators_=_4-4               1.566Ki ± 1%   1.560Ki ± 1%       ~ (p=0.304 n=10)
geomean                                                                                                  1.552Ki        1.551Ki       -0.09%

                                                                                           │ /Users/v/tmp/1before.txt │       /Users/v/tmp/2after.txt       │
                                                                                           │        allocs/op         │ allocs/op   vs base                 │
MergedSeriesChunkRefsSetIterators/with_duplicated_series_=_true_number_of_iterators_=_4-4                  35.00 ± 0%   35.00 ± 0%       ~ (p=1.000 n=10) ¹
MergedSeriesChunkRefsSetIterators/with_duplicated_series_=_false_number_of_iterators_=_4-4                 35.00 ± 0%   35.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                                                    35.00        35.00       +0.00%
¹ all samples are equal
```